Cleaned up moderate_students.py slightly. Reads globals from settings.py. Imports generic make_request function from helpers file instead of defining its own
